feat#18: routing 작업 + 파일 정리
