+++ src/pages/Main/components/Search/components/Categories/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { Box } from "@mui/material";
 import { Global } from "@emotion/react";
-import { useCategory } from "../../../../hook/useCategory";
+import { useCategory } from "../../../../../../hook/useCategory";
 
 const wooyeonCategory = [
   { id: "life", value: "일상" },
