+++ src/pages/Main/components/Search/style.ts
